+++ src/er_c.rs
@@ -60,8 +60,8 @@ impl ErStepC {
 /// * variable number of edges
 /// * targets a connectivity
 /// ## Sampling
-/// * for *simple sampling* look at [```SimpleSample``` trait](./traits/trait.SimpleSample.html)
-/// * for *markov steps* look at [```MarkovChain``` trait](../traits/trait.MarkovChain.html)
+/// * for *simple sampling* look at [```SimpleSample``` trait](./sampling/traits/trait.SimpleSample.html)
+/// * for *markov steps* look at [```MarkovChain``` trait](../sampling/traits/trait.MarkovChain.html)
 /// ## Other
 /// * for topology functions look at [`GenericGraph`](../generic_graph/struct.GenericGraph.html)
 /// * to access underlying topology or manipulate additional data look at [```WithGraph``` trait](../traits/trait.WithGraph.html)