+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 * implemented `Dot` for `GenericGraph`
 * new trait `DotExtra`
 * implemented `DotExtra` for `GenericGraph`
-* `MarkovChain` now has member: `monte_carlo_metropolis<F, G, H, Rng>`
+* new trait `Metropolis`
 
 ## macro
 * add `dot_options!` macro for ease of use